requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via httpcore
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
build==1.2.2.post1
    # via pip-tools
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via pip-tools
cryptography==44.0.2
    # via social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
deprecation==2.1.0
    # via postgrest
django==4.2.10
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   drf-yasg
    #   social-auth-app-django
django-cors-headers==4.3.1
    # via -r requirements.in
django-filter==23.5
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   drf-yasg
drf-yasg==1.21.7
    # via -r requirements.in
ecdsa==0.19.0
    # via python-jose
exceptiongroup==1.2.2
    # via anyio
gotrue==2.9.1
    # via supabase
gunicorn==21.2.0
    # via -r requirements.in
h11==0.14.0
    # via httpcore
h2==4.2.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==0.17.3
    # via httpx
httpx[http2]==0.24.1
    # via
    #   gotrue
    #   postgrest
    #   storage3
    #   supabase
    #   supafunc
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
inflection==0.5.1
    # via drf-yasg
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==24.2
    # via
    #   build
    #   deprecation
    #   drf-yasg
    #   gunicorn
pip-tools==7.3.0
    # via -r requirements.in
postgrest==0.13.2
    # via supabase
psycopg2-binary==2.9.9
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic==2.10.6
    # via
    #   gotrue
    #   postgrest
pydantic-core==2.27.2
    # via pydantic
pyjwt==2.10.1
    # via social-auth-core
pyproject-hooks==1.2.0
    # via build
python-dateutil==2.9.0.post0
    # via
    #   realtime
    #   storage3
python-dotenv==1.0.1
    # via -r requirements.in
python-jose==3.3.0
    # via -r requirements.in
python3-openid==3.2.0
    # via social-auth-core
pytz==2025.1
    # via
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0.2
    # via drf-yasg
realtime==1.0.6
    # via supabase
requests==2.31.0
    # via
    #   -r requirements.in
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==2.0.0
    # via social-auth-core
rsa==4.9
    # via python-jose
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpcore
    #   httpx
social-auth-app-django==5.4.0
    # via -r requirements.in
social-auth-core==4.5.6
    # via social-auth-app-django
sqlparse==0.5.3
    # via django
storage3==0.7.7
    # via supabase
strenum==0.4.15
    # via postgrest
supabase==2.2.0
    # via -r requirements.in
supafunc==0.3.3
    # via supabase
tomli==2.2.1
    # via
    #   build
    #   pip-tools
typing-extensions==4.12.2
    # via
    #   anyio
    #   asgiref
    #   pydantic
    #   pydantic-core
    #   realtime
    #   storage3
uritemplate==4.1.1
    # via drf-yasg
urllib3==2.3.0
    # via requests
websockets==12.0
    # via realtime
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
